requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chainer==7.7.0
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.5
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-compressor==2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
django-sass-processor==0.8.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
filelock==3.0.12
Flask==1.1.2
gensim==3.8.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
libsass==0.20.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pandas==1.1.5
passlib==1.7.4
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
rcssmin==1.0.6
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
scipy==1.6.1
six==1.15.0
smart-open==4.2.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
WTForms==2.3.3
